Footer Design & css & NanumSquare applied
